refactor(ui): split page into Server Component + CalculatorView client component
page.tsx is now a Server Component: reads the inflationMeasure cookie and
renders the static shell (PageWrapper, H1). CalculatorView holds all
client-side state and interactivity. Replaces localStorage with a cookie
so the server can render the correct measure on load, eliminating the
post-hydration flicker.
